Remove classes for entities with no infoboxes
