+++ data/ai_map/quality/entities/entity_counts.json
@@ -1,6 +1,6 @@
 {
     "URI": 1,
-    "abstract": 0.9744324785000388,
+    "abstract": 0.9745099558379174,
     "derivedFrom": 0.9994576586348493,
-    "imageURL": 0.7256527465716278
+    "imageURL": 0.7267374293019292
 }
