+++ data/ai_map/quality/entities/image_status.json
@@ -1,5 +1,5 @@
 {
-    "200": 1278,
-    "404": 190,
-    "429": 7898
+    "200": 603,
+    "404": 93,
+    "429": 8684
 }
