+++ data/ai_map/quality/entities/missing_derived_from.json
@@ -6,9 +6,9 @@
     "entities": [
         "http://dbpedia.org/resource/Glossary_of_contract_bridge_terms",
         "http://dbpedia.org/resource/7_service",
+        "http://dbpedia.org/resource/FPQ-16_PARCS",
         "http://dbpedia.org/resource/IEC_27001",
         "http://dbpedia.org/resource/Test_generation",
-        "http://dbpedia.org/resource/FPQ-16_PARCS",
         "http://dbpedia.org/resource/Startupbootcamp",
         "http://dbpedia.org/resource/IEC_7816"
     ]
